+++ packages/ds/angular/checkbox/checkbox.component.ts
@@ -1,0 +1,12 @@
+import { Component, Input } from '@angular/core';
+
+@Component({
+  selector: 'jds-checkbox',
+  templateUrl: './checkbox.component.html',
+  styleUrls: ['./checkbox.component.css']
+})
+export class CheckboxComponent { 
+  @Input() size?: 'small' | 'large';
+  @Input() ariaLabel?: string;
+  @Input() disabled?: boolean;
+}