+++ packages/ds/angular/src/public-api.ts
@@ -3,14 +3,7 @@
  */
 
 // DIRECTIVES
-export * from './common/button/button.component';
-export * from './common/footer/footer.component';
-export * from './common/header/header.component';
-export * from './common/icon/icon.component';
-export * from './common/jl-common.module';
-export * from './common/link/link.component';
 export * from './stencil-generated/components';
 export * from './stencil-generated/index';
 // PACKAGE MODULE
 export { IRCCModule } from './web-module';
-