+++ sample/src/app/app.component.ts
@@ -1,10 +1,7 @@
 import { Component } from '@angular/core';
 import { Router } from '@angular/router';
-import { IIconConfig } from '@ircc-ca/ds-sdc-angular';
 import { TranslateService } from '@ngx-translate/core';
 import { routes } from './routing/app.routing.module';
-// import { IFieldConfig } from 'packages/ds/angular/src/lib/jl-cl/jl-cl/IFormBase';
-
 @Component({
     selector: 'app-root',
     templateUrl: './app.component.html',
@@ -21,17 +18,14 @@ export class AppComponent {
     //   placeholder: '',
     // } as IFieldConfig;
 
-    alicornIconConfig = {
-        unicode: 'f6b0',
-        fontFamily: 'fa-solid',
-    } as IIconConfig;
+    constructor(public translate: TranslateService, private router: Router) {
+        translate.addLangs(['en', 'fr']);
+        translate.setDefaultLang('en');
+    }
 
-    rocketIconConfig = {
-        unicode: 'e027',
-        fontFamily: 'fa-solid',
-    } as IIconConfig;
-
-    constructor(private translate: TranslateService, private router: Router) {}
+    switchLang(lang: string) {
+        this.translate.use(lang);
+    }
 
     findRoute(dir = 1) {
         const route = this.router.url.replace('/', '');