+++ packages/ds/core/CHANGELOG.md
@@ -3,13 +3,59 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+
+## [0.0.10-beta.13](https://github.com/IRCC-ca/ds-sdc/compare/@ircc-ca/ds-sdc-core@0.0.10-beta.12...@ircc-ca/ds-sdc-core@0.0.10-beta.13) (2022-06-14)
+
 ## [0.0.10](https://github.com/IRCC-ca/ds-sdc/compare/@ircc-ca/ds-sdc-core@0.0.9...@ircc-ca/ds-sdc-core@0.0.10) (2022-05-13)
+
 
 **Note:** Version bump only for package @ircc-ca/ds-sdc-core
 
 
 
 
+
+## [0.0.10-beta.12](https://github.com/IRCC-ca/ds-sdc/compare/@ircc-ca/ds-sdc-core@0.0.10-beta.11...@ircc-ca/ds-sdc-core@0.0.10-beta.12) (2022-06-10)
+
+**Note:** Version bump only for package @ircc-ca/ds-sdc-core
+
+
+
+
+
+## [0.0.10-beta.11](https://github.com/IRCC-ca/ds-sdc/compare/@ircc-ca/ds-sdc-core@0.0.10-beta.10...@ircc-ca/ds-sdc-core@0.0.10-beta.11) (2022-06-10)
+
+**Note:** Version bump only for package @ircc-ca/ds-sdc-core
+
+
+
+
+
+## [0.0.10-beta.10](https://github.com/IRCC-ca/ds-sdc/compare/@ircc-ca/ds-sdc-core@0.0.10-beta.9...@ircc-ca/ds-sdc-core@0.0.10-beta.10) (2022-06-08)
+
+**Note:** Version bump only for package @ircc-ca/ds-sdc-core
+
+
+
+
+
+## [0.0.10-beta.9](https://github.com/IRCC-ca/ds-sdc/compare/@ircc-ca/ds-sdc-core@0.0.10-beta.8...@ircc-ca/ds-sdc-core@0.0.10-beta.9) (2022-05-27)
+
+**Note:** Version bump only for package @ircc-ca/ds-sdc-core
+
+
+
+
+
+## [0.0.10-beta.8](https://github.com/IRCC-ca/ds-sdc/compare/@ircc-ca/ds-sdc-core@0.0.10-beta.7...@ircc-ca/ds-sdc-core@0.0.10-beta.8) (2022-05-27)
+
+**Note:** Version bump only for package @ircc-ca/ds-sdc-core
+
+
+
+
+
+## [0.0.10-beta.7](https://github.com/IRCC-ca/ds-sdc/compare/@ircc-ca/ds-sdc-core@0.0.10-beta.6...@ircc-ca/ds-sdc-core@0.0.10-beta.7) (2022-05-20)
 
 ## [0.0.10-beta.0](https://github.com/IRCC-ca/ds-sdc/compare/@ircc-ca/ds-sdc-core@0.0.9...@ircc-ca/ds-sdc-core@0.0.10-beta.0) (2022-05-13)
 
@@ -18,6 +64,52 @@ See [Conventional Commits](https://conventionalcommits.org) for commit guideline
 
 
 
+
+## [0.0.10-beta.6](https://github.com/IRCC-ca/ds-sdc/compare/@ircc-ca/ds-sdc-core@0.0.10-beta.5...@ircc-ca/ds-sdc-core@0.0.10-beta.6) (2022-05-19)
+
+**Note:** Version bump only for package @ircc-ca/ds-sdc-core
+
+
+
+
+
+## [0.0.10-beta.5](https://github.com/IRCC-ca/ds-sdc/compare/@ircc-ca/ds-sdc-core@0.0.10-beta.4...@ircc-ca/ds-sdc-core@0.0.10-beta.5) (2022-05-19)
+
+**Note:** Version bump only for package @ircc-ca/ds-sdc-core
+
+
+
+
+
+## [0.0.10-beta.4](https://github.com/IRCC-ca/ds-sdc/compare/@ircc-ca/ds-sdc-core@0.0.10-beta.3...@ircc-ca/ds-sdc-core@0.0.10-beta.4) (2022-05-16)
+
+**Note:** Version bump only for package @ircc-ca/ds-sdc-core
+
+## [0.0.10-beta.3](https://github.com/IRCC-ca/ds-sdc/compare/@ircc-ca/ds-sdc-core@0.0.10-beta.2...@ircc-ca/ds-sdc-core@0.0.10-beta.3) (2022-05-16)
+
+**Note:** Version bump only for package @ircc-ca/ds-sdc-core
+
+## [0.0.10-beta.2](https://github.com/IRCC-ca/ds-sdc/compare/@ircc-ca/ds-sdc-core@0.0.10-beta.1...@ircc-ca/ds-sdc-core@0.0.10-beta.2) (2022-05-16)
+
+**Note:** Version bump only for package @ircc-ca/ds-sdc-core
+
+## [0.0.9-beta.0](https://github.com/IRCC-ca/ds-sdc/compare/@ircc-ca/ds-sdc-core@0.0.8...@ircc-ca/ds-sdc-core@0.0.9-beta.0) (2022-05-06)
+
+**Note:** Version bump only for package @ircc-ca/ds-sdc-core
+
+## [0.0.8](https://github.com/IRCC-ca/ds-sdc/compare/@ircc-ca/ds-sdc-core@0.0.7...@ircc-ca/ds-sdc-core@0.0.8) (2022-05-05)
+
+## [0.0.10-beta.2](https://github.com/IRCC-ca/ds-sdc/compare/@ircc-ca/ds-sdc-core@0.0.10-beta.1...@ircc-ca/ds-sdc-core@0.0.10-beta.2) (2022-05-16)
+
+**Note:** Version bump only for package @ircc-ca/ds-sdc-core
+
+## [0.0.10-beta.1](https://github.com/IRCC-ca/ds-sdc/compare/@ircc-ca/ds-sdc-core@0.0.9...@ircc-ca/ds-sdc-core@0.0.10-beta.1) (2022-05-13)
+
+**Note:** Version bump only for package @ircc-ca/ds-sdc-core
+
+## [0.0.10-beta.0](https://github.com/IRCC-ca/ds-sdc/compare/@ircc-ca/ds-sdc-core@0.0.9...@ircc-ca/ds-sdc-core@0.0.10-beta.0) (2022-05-13)
+
+**Note:** Version bump only for package @ircc-ca/ds-sdc-core
 
 ## [0.0.9](https://github.com/IRCC-ca/ds-sdc/compare/@ircc-ca/ds-sdc-core@0.0.8...@ircc-ca/ds-sdc-core@0.0.9) (2022-05-12)