+++ packages/ds/web/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.0.5-beta.2](https://github.com/IRCC-ca/ds-sdc/compare/@ircc-ca/ds-sdc-web@0.0.5-beta.1...@ircc-ca/ds-sdc-web@0.0.5-beta.2) (2022-05-24)
+
+**Note:** Version bump only for package @ircc-ca/ds-sdc-web
+
+
+
+
+
+## [0.0.5-beta.1](https://github.com/IRCC-ca/ds-sdc/compare/@ircc-ca/ds-sdc-web@0.0.5-beta.0...@ircc-ca/ds-sdc-web@0.0.5-beta.1) (2022-05-19)
+
+**Note:** Version bump only for package @ircc-ca/ds-sdc-web
+
+
+
+
+
+## [0.0.5-beta.0](https://github.com/IRCC-ca/ds-sdc/compare/@ircc-ca/ds-sdc-web@0.0.4...@ircc-ca/ds-sdc-web@0.0.5-beta.0) (2022-05-13)
+
+**Note:** Version bump only for package @ircc-ca/ds-sdc-web
+
+
+
+
+
 ## [0.0.4](https://github.com/IRCC-ca/ds-sdc/compare/@ircc-ca/ds-sdc-web@0.0.3...@ircc-ca/ds-sdc-web@0.0.4) (2022-05-05)
 
 **Note:** Version bump only for package @ircc-ca/ds-sdc-web